chore: update ESLint to v10
- Update eslint, @eslint/js, and typescript-eslint
- Convert eslint config from CJS to ESM
- Remove @types/eslint__js (now bundled)
- Fix new no-useless-assignment errors
