+++ package.json
@@ -19,21 +19,20 @@
     "@commitlint/cli": "20.5.0",
     "@commitlint/config-conventional": "20.5.0",
     "@eslint/compat": "2.0.3",
-    "@eslint/js": "9.39.4",
+    "@eslint/js": "^10.0.1",
     "@sveltejs/adapter-static": "3.0.10",
     "@sveltejs/kit": "2.55.0",
     "@sveltejs/vite-plugin-svelte": "3.1.2",
     "@tailwindcss/aspect-ratio": "0.4.2",
     "@tailwindcss/forms": "0.5.11",
     "@types/balanced-match": "3.0.2",
-    "@types/eslint__js": "9.14.0",
     "@types/node": "24.12.0",
     "@types/path-browserify": "1.0.3",
     "@types/ua-parser-js": "0.7.39",
     "@types/webappsec-credential-management": "0.6.9",
     "@types/wicg-file-system-access": "2023.10.7",
     "autoprefixer": "10.4.27",
-    "eslint": "9.39.4",
+    "eslint": "^10.1.0",
     "eslint-config-prettier": "10.1.8",
     "eslint-import-resolver-typescript": "4.4.4",
     "eslint-plugin-headers": "1.3.4",
@@ -54,7 +53,7 @@
     "tailwindcss": "3.4.19",
     "tslib": "2.8.1",
     "typescript": "5.9.3",
-    "typescript-eslint": "8.25.0",
+    "typescript-eslint": "^8.57.2",
     "vite": "5.4.21"
   },
   "dependencies": {
